requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
Werkzeug==2.3.7
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.0 